+++ ntex-io/CHANGES.md
@@ -6,6 +6,8 @@
 
 * Better Io/IoRef api separation
 
+* DefaultFilter renamed to Base
+
 ## [0.1.0-b.1] - 2021-12-19
 
 * Remove ReadFilter/WriteFilter traits.